refactor: remove dead code and unused files
- Removed empty simple-llm-evaluator.ts file
- Removed unused security-response-parser.ts, base-response-parser.ts, structured-output-schemas.ts files
- Removed unused batch-schemas.ts from types directory
- Removed unused NewLLMEvaluationResultSchema and related types from enhanced-security.ts
- Removed unused NewEvaluationResultSchema and FinalEvaluationResultSchema
- Cleaned up type definitions to keep only actively used schemas
- Maintained all functional code while removing dead code
- All builds and lint checks pass successfully
